[367. Valid Perfect Square][Accepted]committed by LeetCode Extension

+++ 367-Valid-Perfect-Square/solution.java
@@ -1,11 +1,11 @@
 public class Solution {
     public boolean isPerfectSquare(int num) {
         if(num < 0) return false;
-        int low = 0, high = num;
+        long low = 0, high = num;
         while(low <= high) {
-            int mid = low + (high - low) / 2;
-            if(mid * mid == num)    return true;
-            if(mid * mid > num) high = mid - 1;
+            long mid = low + (high - low) / 2;
+            if(mid * mid == (long)num)    return true;
+            if(mid * mid > (long)num) high = mid - 1;
             else low = mid + 1;
         }
         return false;
